+++ CHANGELOG.rst
@@ -29,6 +29,11 @@ Removed
 Security
 --------
 
+`v0.6`_ (29, Nov., 2019)
+==========================
+
+* same as v0.6b2
+
 `v0.6b2`_ (24, Nov., 2019)
 ==========================
 
@@ -322,7 +327,8 @@ Fixed
 .. _requests: https://pypi.org/project/requests
 .. _argparse: https://pypi.org/project/argparse/
 
-.. _Unreleased: https://github.com/miurahr/aqtinstall/compare/v0.6b2...HEAD
+.. _Unreleased: https://github.com/miurahr/aqtinstall/compare/v0.6...HEAD
+.. _v0.6: https://github.com/miurahr/aqtinstall/compare/v0.6b2...v0.6
 .. _v0.6b2: https://github.com/miurahr/aqtinstall/compare/v0.6b1...v0.6b2
 .. _v0.6b1: https://github.com/miurahr/aqtinstall/compare/v0.6a2...v0.6b1
 .. _v0.6a2: https://github.com/miurahr/aqtinstall/compare/v0.6a1...v0.6a2